🛡️ Sentinel: [security enhancement] Enhance environment variable redaction
- Expanded `SENSITIVE_VAR_PATTERNS` in `src/lib/security/env-validation.ts` with 22 new patterns.
- Merged tests into `tests/security/env-validation.test.ts` for better CI compatibility.
- Updated security journal in `.jules/sentinel.md`.
